+++ packages/ice/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.0.1
 
 - [feat] support external config for server
+- [feat] redirect imports for data loader
 
 ## v3.0.0